// src/modules/settings/types/user-address.ts
export type UserProfileAddress = {
	id: number
	first_name: string
	last_name: string
	email: string
	phone_number: string
	fax_number: string
	address: string
	city: string
	is_default: boolean
}

export type UserProfileAddressPayload = {
	first_name: string
	last_name: string
	email: string
	phone_number: string
	fax_number: string
	address: string
	city: string
	is_default: boolean
}